requirements.txt
flask==3.0.2
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-mail==0.9.1
python-dotenv==1.0.1
PyJWT==2.8.0
cryptography==42.0.5
python-magic-bin
Werkzeug==3.0.1